threatintel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secnews"
version = "1.1.0"
description = "Security News Aggregator & Digest — a CLI tool for daily threat intel briefings"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["security", "threat-intel", "CVE", "news", "cli", "cybersecurity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "feedparser>=6.0",
    "requests>=2.31",
    "rapidfuzz>=3.6",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.scripts]
secnews = "secnews.cli.main:main"

[project.optional-dependencies]
ai = [
    "anthropic>=0.40",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "responses>=0.25",
]
all = [
    "anthropic>=0.40",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "responses>=0.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["secnews*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
